requirements.txt
# YOLOv11
ultralytics>=8.2.103

# Procesamiento de imágenes y video
opencv-python==4.8.1.78
pillow==10.2.0

# Ciencia de datos
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2

# Webapp para demos
streamlit==1.28.2

# Gestión de datasets
roboflow==1.1.25
supervision==0.16.0
kaggle==1.5.16
pyyaml==6.0.1